Added centralized authentication status hook and updated navigation logic for admin custom user:
AuthStatus.js:
- Created a new useAuthStatus hook to centralize logic for checking if a user is authenticated and whether they are an admin.
- Utilized axiosReq to check the user admin status and stored it in state.
- Returned isAdmin and isAuthenticated to be used across components.

Menu.js:
- Replaced local authentication and admin checks with useAuthStatus hook.
- Updated the booking button to navigate based on the user’s role (/admin/bookings for admin and /bookings for regular users).
- Refactored handleBookAppointmentClick and removed redundant authentication checks.

NavBar.js:
- Removed local admin state and replaced authentication and admin checks with the useAuthStatus hook.
- Adjusted the "Booking" link to conditionally navigate to the appropriate page based on the user’s role (admin or regular user).
- Refactored handleLogout to redirect users to the home page after logging out.
